// src/app/shared/enum/score-grade.enum.ts
export enum NutriscoreGrade {
  A = 'a',
  B = 'b',
  C = 'c',
  D = 'd',
  E = 'e',
  UNKNOWN = 'unknown',
}

export enum EcoscoreGrade {
  A = 'a',
  B = 'b',
  C = 'c',
  D = 'd',
  UNKNOWN = 'unknown',
}

export enum NovagroupGrade {
  GRADE_1 = 1,
  GRADE_2 = 2,
  GRADE_3 = 3,
  GRADE_4 = 4,
  UNKNOWN = 0,
}